+++ CHANGELOG.md
@@ -6,8 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
-In progress…
+### Added
+- Support passing Node resources in JS functions
 
 ## [1.0.2] - 2018-06-18
 ### Fixed